Enable filtering of results in 'Manage Data'
Allows the user to filter the datasets that they have access to.  For
drafts this is done by searching inside the title for the filter term,
but for datasets it uses package_search.

Although the default sort order is most recently modified first, I'm not
sure this makes sense once applying a filter as where datasets are
coming from CKAN it is sometimes not obvious why the result is in the
list (it may have matched on description).  It may be best to constrain
the filter to only search CKAN titles rather than any field.