Enable skipping of tasks that the user is not interested in
Provides a URL /task/skip/ID so that the current user can skip a task
that they are not interested in without the task being deleted (so that
other members of the organisation can still see it).

Changes the url in dashboard.html so that clicking the skip link takes
the user to the skip view, and then redirects them back to the
dashboard.