Move slug handling to the form so we can handle errors on bad slugs
If we get something that can't be slugified we can now complain that the
form isn't valid and show an error.  We are asserting that it is a bad
title so when we get around to error messages defined in forms we need
to make sure we log it as a bad title.

Fixes #69