Asks the user to choose which organisation to publish under.
Some users are in multiple organisations, so that we can progress to
actually publishing in CKAN (where we need an organisation) then we will
simply ask the user to choose which organisation after they have
provided the title/description.  This currently doesn't get skipped if
the user is only in one organisation (this will be done in a separate
PR).

Requires a migration, and is also likely to break front-end tests.

THis is expected to be re-worked, or moved or both, but is in place
simply to avoid blocking other work.