Resolves an error where trying to publish pushed user back to organisation
Fixes an error where is you randomly go to the check_dataset page and
try to publish (like after an edit) when user was redirected back to
organisation page.  This is because we store the data in the draft
object, rather than relying on the form data in the session.

Also, now deletes draft after successful publish.