Remove the wizard so that views are clear and explicit
The Wizard View introduced lots of problems with our custom flow that
were difficult and messy to work around, made worse because the wizard
was pretty opaque.

This PR changes the views to be single function views so that the path
through the flow is explicit and easier to work with.  This does
introduce repetition of code but we can refactor this some other time
(see drafts.views._redirect_to for example).

Removes all of the wizard stuff from templates, and also adds a View
link in manage data that just opens a CKAN window so we can test (for
now) what the end result looks like.

Editing from /manage will now redirect the user to check_dataset, and if
it is a published dataset will also copy the CKAN data into a local
Draft.  This means it is possible to have the same dataset twice, once
as published and once as Draft (which is the current dataset with
changes) which we will need to make a decision on how it works.