Auto select org when user only a member of one
When a user is only in one organisation, there is no need to ask them
which organisation they are publishing for.  This PR will auto-select
the single org if they are only in one.

As a result, on the check_dataset page we don't offer them the option of
changing the organisation.