Edit full dataset endpoint.
Adds a new url for editing the entire dataset (at once) and uses a new
form that includes all of the relevant fields EXCEPT files.
Files/links/whatever will have to be handled as a sub-form and we still
need to decide whether we are checking URLs on save, or just new URLs,
or neither.