Allows us to change app-specific config at runtime w/o redeploy
For some configuration options, for instance the CKAN that we want to
publish to (for alpha only), this allows us to change the property at
runtime (via the admin) without redeploying the whole app.

It will also allow us to disable certain properties as a mechanism for
disabling functionality during Alpha.  e.g is we deactivate the ckan
property we can stop the publishing actually sending data to CKAN.