Change dataset IDs to a UUID.
URLS in existing DGU support /dataset/name and /dataset/id and so to
make sure we can maintain these links in Alpha, this switches over to
using UUID instead of int for ids.

This will break if you have existing data, so best to delete the
contents of the datasets_dataset table.