Allows flow state to be stored in the session.
Uses a template tag in the templates for determining if the current
request is editing or checking and allows for flow to be changed as a
result.  Typically this will look like

```
  {% load flow_state %}

  {% if request|is_editing %}
     Editing
  {% elif request|is_checking %}
     Checking
  {% else %}
     Neither
  {% endif %}

```

Each view function now includes a call to _set_flow_state which will set
the current state based on the presence of a state= query parameter.  If
it is missing then no change it made to the state value.