Refuse to publish if there are missing fields.
As we require some fields to be present, but want to allow skipping
through the main flow this PR separates out the presence validation
(does this field have a value at all) and allow us to show the errors on
the check_dataset page.

Have added some HTML to check_dataset.html but it is not intended for
actual use, rather just to show how to process the validation errors.