Initial attempt at importing DGU datasets
The dataset_fixtures script can be used to generate the datasets.json
and datafiles.json.  This script is still incomplete and probably
inaccurate, but it's a start.

Because there are so many changes to the DB, it is probably better to
reset and start with a clean one.  This is mostly to allow import of the
existing data.

```
    ./manage.py loaddata datasets
    ./manage.py loaddata datafiles
```