A management command for checking links.
As a temporary measure provides a management command for checking
whether, for a specific organisation, any datasets contain broken links.

Once a dataset is identified as containing broken links, a new task is
created for the dataset and the datafiles are marked as broken.

Contains a migration, and to be safe it is best to run 'delete from
tasks_task;' on the database before use'