Show non-organisation, but owned datasets in manage_data
As organisation will soon be optional, we need to support showing drafts
datasets that may not have an organisation attached.  For this we need
to make sure that the dataset list shows datasets when the user is a
member, OR where the user is the creator (which makes it a personal
draft, rather than a team draft).

Fixes #267