Use messages framework to tell user about deletions
Rather than passing ?result=deleted to manage_data, we instead use the
messages framework to set the text in the view.  See delete_dataset if
we need to pass arguments to the i18n.

We use extra_tags with the dataset name if we want a link to view it.