Add owners to datasets
An owner of a dataset is distinct from the creator.  It might change
over time but we may still want to know who created it.

This would allow assigning datasets to people (if this was a thing we
might want) but is primarily for ensuring 'somebody' owns it even if the
creator stops using the system.