Allow dataset_list to return subsets of the objects
Fetching model objects is expensive, and in cases where we want to
filter and return a couple of fields to the browser is inefficient.

This PR, apart from adding some tests for dataset_list also allow you to
specify the fields you want returned from the lookup.  If you pass a
list of field names in ```fields``` you will receive back a QuerySet
that contains a list of dicts, containing only those fields.  You can
then cast this to list and return to the client.

Fixes #443