Preparation for determining overdue datasets
Looks for the start date of the most recent datafile in a dataset.  Then
checks whether it is more than 'time' days old, where 'time' is the
appropriate number of days for the frequency.  As well as overdue, we
may eventually want to be able to determine how close it is to overdue
for notifications.